refactor(web): extract SlidingTabs from earn chart pane
